+++ src/Hash/hashEncadeamentoExplicito.cpp
@@ -5,13 +5,14 @@
 
 using namespace HashEncadeamentoExplicitoClass;
 
-int funcaoHashingPosicao(int valorDoNode, int tamanhoDoArquivo)
-{
+int funcaoHashingPosicao(long int valorDoNode,
+    long int tamanhoDoArquivo) {
     return valorDoNode % tamanhoDoArquivo;
 }
 
-int buscaPosicaoInicial(Node *arquivo, int valorNovoNode, int tamanhoDoArquivo)
-{
+int buscaPosicaoInicial(Node *arquivo,
+    long int valorNovoNode,
+    long int tamanhoDoArquivo) {
     int posicaoInicial = funcaoHashingPosicao(valorNovoNode, tamanhoDoArquivo);
     return posicaoInicial;
 }
@@ -19,10 +20,11 @@ int buscaPosicaoInicial(Node *arquivo, int valorNovoNode, int tamanhoDoArquivo)
 /* 
 * Encontra o ultimo nó da lista encadeada de forma a apontar para o próximo nó
 */
-int encontrarUltimoNoLivreNaListaEncadeada(Node* arquivo, int& posicaoControle,
-    int& quantidadeDeAcessosTotalHEEAE, int posicaoInicialDeInsercaoNoArquivo) {
+int encontrarUltimoNoLivreNaListaEncadeada(Node* arquivo, long int& posicaoControle,
+    long int& quantidadeDeAcessosTotalHEEAE,
+    long int posicaoInicialDeInsercaoNoArquivo) {
 
-    int posicaoAtualNaLista = posicaoInicialDeInsercaoNoArquivo;
+    long int posicaoAtualNaLista = posicaoInicialDeInsercaoNoArquivo;
     while (arquivo[posicaoAtualNaLista].getProximoNode() != -1) {
         posicaoAtualNaLista = arquivo[posicaoAtualNaLista].getProximoNode();
         quantidadeDeAcessosTotalHEEAE++;
@@ -34,18 +36,21 @@ int encontrarUltimoNoLivreNaListaEncadeada(Node* arquivo, int& posicaoControle,
 * Encontra o primeiro nó livre no arquivo, de cima para baixo, verificando se a posição está ocupada
 * para guardar a posição que deve marcar o próxima posição que será inserida caso haja colisão
 */
-void encontrarPrimeiroNodeLivre(Node* arquivo, int& posicaoControle, int& quantidadeDeAcessosTotalHEEAE) {
+void encontrarPrimeiroNodeLivre(Node* arquivo, 
+    long int& posicaoControle,
+    long int& quantidadeDeAcessosTotalHEEAE) {
     while (posicaoControle >= 0 && arquivo[posicaoControle].isEstaOcupado()) {
         posicaoControle--;
         quantidadeDeAcessosTotalHEEAE++;
     }
 }
 
-void HashEncadeamentoExplicito::insercaoNode(Node* arquivo, int valorNovoNode,
-                                              int tamanhoDoArquivo,
-                                              int& posicaoControle,
-                                              int& quantidadeDeAcessosTotalHEEAE) {
-    int posicaoInicialDeInsercaoNoArquivo = buscaPosicaoInicial(arquivo, valorNovoNode, tamanhoDoArquivo);
+void HashEncadeamentoExplicito::insercaoNode(Node* arquivo,
+    long int valorNovoNode,
+    long int tamanhoDoArquivo,
+    long int& posicaoControle,
+    long int& quantidadeDeAcessosTotalHEEAE) {
+    long int posicaoInicialDeInsercaoNoArquivo = buscaPosicaoInicial(arquivo, valorNovoNode, tamanhoDoArquivo);
 
     if (!arquivo[posicaoInicialDeInsercaoNoArquivo].isEstaOcupado())
     {
@@ -53,7 +58,7 @@ void HashEncadeamentoExplicito::insercaoNode(Node* arquivo, int valorNovoNode,
         arquivo[posicaoInicialDeInsercaoNoArquivo].setValorDoNode(valorNovoNode);
         arquivo[posicaoInicialDeInsercaoNoArquivo].setEstaOcupado(true);
     } else {
-        int posicaoAtualNaLista = encontrarUltimoNoLivreNaListaEncadeada(arquivo, posicaoControle,
+        long int posicaoAtualNaLista = encontrarUltimoNoLivreNaListaEncadeada(arquivo, posicaoControle,
                                 quantidadeDeAcessosTotalHEEAE, posicaoInicialDeInsercaoNoArquivo);
     
         encontrarPrimeiroNodeLivre(arquivo, posicaoControle, quantidadeDeAcessosTotalHEEAE);
