+++ src/Hash/hashLinearProbing.cpp
@@ -13,21 +13,25 @@ int hashLinearProbing1(int chave, int quantidade_de_registros) {
 }
 
 // Função para inserção das chaves usando hashing com encadeamento linear
-void HashLinearProbing::insertLinearProbingHash(vector<int>& tabela_hash, int chave,
-        int quantidade_de_registros, int& total_chaves_inseridas, int& total_acessos) {
+void HashLinearProbing::insertLinearProbingHash(
+    vector<long int>& tabela_hash,
+    long int chave,
+    long int quantidade_de_registros,
+    long int& total_chaves_inseridas,
+    long int& total_acessos) {
     // while(total_chaves_inseridas < quantidade_de_registros) { // se o total de chaves alcançar o número de registros disponíveis encerra o loop
         
         if(chave == 0) return; //se chave for 0, saia do loop
 
-        int h1 = hashLinearProbing1(chave, quantidade_de_registros); // chamada para calcular o h1
+        long int h1 = hashLinearProbing1(chave, quantidade_de_registros); // chamada para calcular o h1
 
         if (tabela_hash[h1] == -1) { // caso em que, usando h1, a posição calculada está disponível
             tabela_hash[h1] = chave; // chave inserida 
             total_chaves_inseridas++; // incrementa o total de chaves inseridas
             total_acessos++; // incrementa o total de acessos
         } else { // caso em que, usando h1, a posição calculada está ocupada
-            int i = 1;
-            int nova_posicao = (h1 + i) % quantidade_de_registros;
+            long int i = 1;
+            long int nova_posicao = (h1 + i) % quantidade_de_registros;
             total_acessos++; // incrementa o total de acessos
 
             while(tabela_hash[nova_posicao] != -1) { /* o programa continua até que encontre uma posição que esteja disponível,
