+++ src/Hash/hashPerfect.cpp
@@ -6,14 +6,14 @@
 using namespace HashPerfectClass;
 
 //Função para criação da tabela do nível 1
-vector<vector<int>> criar_tabela_nivel_um(int quantidade_de_registros){
-    vector<vector<int>> tabela_nivel_um(quantidade_de_registros);
+vector<vector<long int>> criar_tabela_nivel_um(long int quantidade_de_registros) {
+    vector<vector<long int>> tabela_nivel_um(quantidade_de_registros);
     return tabela_nivel_um;
 }
 
 //Função para criação da tabela do nível 2
-vector<vector<int>> criar_tabela_nivel_dois(int quantidade_de_registros){
-    vector<vector<int>> tabela_nivel_dois(quantidade_de_registros);
+vector<vector<long int>> criar_tabela_nivel_dois(long int quantidade_de_registros) {
+    vector<vector<long int>> tabela_nivel_dois(quantidade_de_registros);
     return tabela_nivel_dois;
 }
 
@@ -24,8 +24,13 @@ void gerar_variaveis_aleatorias(int &a, int &b) {
     b = rand() % 101;
 }
 
-void HashPerfect::insertPerfectHash(vector<vector<int>>& nivel_um, vector<vector<int>>& nivel_dois,
-    int chave, int quantidade_de_registros, int a, int b, int& total_acessos, int& total_chaves_inseridas) {
+void HashPerfect::insertPerfectHash(vector<vector<long int>>& nivel_um,
+                vector<vector<long int>>& nivel_dois,
+                long int chave,
+                long int quantidade_de_registros,
+                int a, int b,
+                long int& total_acessos,
+                long int& total_chaves_inseridas) {
     
     vector<int> chaves; // Um vetor de chaves, auxiliar, é usado para armazenar as chaves, passadas na entrada, que serão inseridas nos registros
 
@@ -38,7 +43,7 @@ void HashPerfect::insertPerfectHash(vector<vector<int>>& nivel_um, vector<vector
 
     // Inserção na tabela principal
     for (unsigned long int i = 0; i < chaves.size(); i++) {
-        int h = ((a * chaves[i] + b) % 101) % quantidade_de_registros; // cálculo do hash
+        long int h = ((a * chaves[i] + b) % 101) % quantidade_de_registros; // cálculo do hash
         nivel_um[h].resize(nivel_um[h].size() + 1, chaves[i]); // insere a chave no nível um
         total_acessos++; // incrementa o total de acessos
         total_chaves_inseridas++; // incrementa o total de chaves inseridas
@@ -48,21 +53,21 @@ void HashPerfect::insertPerfectHash(vector<vector<int>>& nivel_um, vector<vector
     na tabela secundária tem tamanho igual a um referente a única chave a ser inserida. Por outro lado, quando em uma dada
     posição na tabela principal há mais de uma chave, essa posição na tabela secundária tem tamanho igual ao número de chaves
     nessa posição elevado ao quadrado*/
-    for (int i = 0; i < quantidade_de_registros; i++) {
-        int n = nivel_um[i].size();
+    for (long int i = 0; i < quantidade_de_registros; i++) {
+        long int n = nivel_um[i].size();
         if (n == 1) {
-            int h = ((a * nivel_um[i][0] + b) % 101) % quantidade_de_registros;
+            long int h = ((a * nivel_um[i][0] + b) % 101) % quantidade_de_registros;
             nivel_dois[h].resize(1, nivel_um[i][0]);
             total_acessos++; // incrementa o total de acessos
         } else if (n > 1) {
             nivel_dois[i].resize(n * n);
-            for (int j = 0; j < n; j++) {
-                int h = ((a * nivel_um[i][j] + b) % 101) % (n * n);
+            for (long int j = 0; j < n; j++) {
+                long int h = ((a * nivel_um[i][j] + b) % 101) % (n * n);
                 if (nivel_dois[i][h] == 0) {
                     nivel_dois[i][h] = nivel_um[i][j];
                     total_acessos++; // incrementa o total de acessos
                 } else {
-                    int k = h; // variável k, variável auxiliar que recebe o valor de hash
+                    long int k = h; // variável k, variável auxiliar que recebe o valor de hash
                     do {
                         k = (k + 1) % (n * n); // atualiza o valor de hash até que aja posição buscada disponível
                         total_acessos++; // incrementa o total de acessos
